microservices/classification/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27
pika==1.3.2
openai==1.3.5
python-dotenv==1.0.0
